pyproject.toml
[project]
name = "rickroll-firewall"
version = "0.1.0"
description = "rickroll-firewall"
authors = [
    {name = "Zahcoder34",email = "Zakhriw@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "ascender-framework (>=2.0.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
